pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.0.1"
description = ""
authors = ["Kirill Lapushinskiy <kirilllapushinskiy@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
python-dotenv = "^0.20.0"
marshmallow = "^3.17.0"
aiohttp-apispec = "^2.2.3"
SQLAlchemy = "^1.4.39"
asyncpg = "^0.26.0"
alembic = "^1.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
